services/compliance/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
httpx==0.25.2
pydantic==2.5.0
